[#1820] Add validation for status
