+++ src/openzaak/components/zaken/api/serializers/zaken.py
@@ -34,6 +34,7 @@ from vng_api_common.utils import get_help_text
 from vng_api_common.validators import IsImmutableValidator, UntilNowValidator
 
 from openzaak.components.documenten.api.fields import EnkelvoudigInformatieObjectField
+from openzaak.components.zaken.validators import CorrectZaaktypeValidator
 from openzaak.contrib.verzoeken.validators import verzoek_validator
 from openzaak.utils.api import (
     create_remote_objectcontactmoment,
@@ -71,7 +72,6 @@ from ...models import (
     ZaakVerzoek,
 )
 from ..validators import (
-    CorrectZaaktypeValidator,
     DateNotInFutureValidator,
     EndStatusIOsIndicatieGebruiksrechtValidator,
     EndStatusIOsUnlockedValidator,
